+++ src/__stories__/header-story.tsx
@@ -6,10 +6,9 @@ import {
     NavigationBreadcrumbs,
     ResponsiveLayout,
     Placeholder,
-    Text1,
-    Text2,
+    Callout,
+    IconInformationUserLight,
 } from '..';
-import {useTextField, useCheckbox} from './helpers';
 
 export default {
     title: 'Components/Headers/Header',
@@ -18,111 +17,106 @@ export default {
     },
 };
 
-export const Default: StoryComponent = () => {
-    const [pretitle, pretitleTextField] = useTextField('Pretitle', 'Your last bill');
-    const [title, titleTextField] = useTextField('Title', 'December bill is now available');
-    const [description, descriptionTextField] = useTextField('Description', 'This is a description');
-    const [isInverse, inverseCheckbox] = useCheckbox('Inverse', true);
-    const [withExtraContent, extraContentCheckbox] = useCheckbox('With extra content', true);
-    const [extraSideBySide, extraSideBySideCheckbox] = useCheckbox(
-        'Extra content placed on the right in desktop',
-        true
-    );
-    const [withBreadcrumbs, breadcrumbsCheckbox] = useCheckbox('With breadcrumbs (desktop only)', true);
-    const [bleed, bleedCheckbox] = useCheckbox('Bleed', false);
+type Args = {
+    withHeader: boolean;
+    pretitle: string;
+    title: string;
+    description: string;
+    small: boolean;
+    isInverse: boolean;
+    withExtraContent: boolean;
+    sideBySideExtraOnDesktop: boolean;
+    withBreadcrumbs: boolean;
+    bleed: boolean;
+    noPaddingY: boolean;
+};
+
+export const Default: StoryComponent<Args> = ({
+    withHeader,
+    isInverse,
+    bleed,
+    sideBySideExtraOnDesktop,
+    withBreadcrumbs,
+    pretitle,
+    title,
+    description,
+    small,
+    withExtraContent,
+    noPaddingY,
+}) => {
     return (
         <Stack space={16}>
-            <div data-testid="header-layout">
-                <HeaderLayout
-                    isInverse={isInverse}
-                    bleed={bleed}
-                    sideBySideExtraOnDesktop={extraSideBySide}
-                    breadcrumbs={
-                        withBreadcrumbs ? (
-                            <NavigationBreadcrumbs
-                                title="Bills"
-                                breadcrumbs={[{title: 'Account', url: '/consumptions'}]}
-                            />
-                        ) : undefined
-                    }
-                    header={<Header pretitle={pretitle} title={title} description={description} />}
-                    extra={withExtraContent ? <Placeholder /> : undefined}
-                />
-            </div>
+            <HeaderLayout
+                dataAttributes={{testid: 'header-layout'}}
+                isInverse={isInverse}
+                bleed={bleed}
+                sideBySideExtraOnDesktop={sideBySideExtraOnDesktop}
+                noPaddingY={noPaddingY}
+                breadcrumbs={
+                    withBreadcrumbs ? (
+                        <NavigationBreadcrumbs
+                            title="Bills"
+                            breadcrumbs={[{title: 'Account', url: '/consumptions'}]}
+                        />
+                    ) : undefined
+                }
+                header={
+                    withHeader ? (
+                        <Header pretitle={pretitle} title={title} description={description} small={small} />
+                    ) : undefined
+                }
+                extra={withExtraContent ? <Placeholder /> : undefined}
+            />
             <ResponsiveLayout>
-                <Stack space={16}>
-                    {breadcrumbsCheckbox}
-                    {pretitleTextField}
-                    {titleTextField}
-                    {descriptionTextField}
-                    {bleedCheckbox}
-                    {inverseCheckbox}
-                    {extraContentCheckbox}
-                    {extraSideBySideCheckbox}
-                </Stack>
+                <Callout
+                    icon={<IconInformationUserLight />}
+                    title="Note:"
+                    description="A HeaderLayout can be rendered without a Header (in webviews) to allow delegating the
+                        header visualization to the native app. For example, in Novum App (Mi Movistar, Meu
+                        Vivo, etc), the Start tab's greeting is rendered nativelly in the apps."
+                />
             </ResponsiveLayout>
         </Stack>
     );
 };
 
 Default.storyName = 'Header';
-
-/**
- * The header is optional in order to allow webviews to delegate the header visualization to the surrounding native app.
- * For example, in Novum App, the Start tab's greeting is rendered nativelly in the apps and via web in desktop.
- */
-export const NoHeader: StoryComponent = () => {
-    const [extraSideBySide, extraSideBySideCheckbox] = useCheckbox(
-        'Extra content placed on the right in desktop',
-        true
-    );
-    return (
-        <Stack space={16}>
-            <div data-testid="header-layout">
-                <HeaderLayout
-                    isInverse
-                    sideBySideExtraOnDesktop={extraSideBySide}
-                    breadcrumbs={
-                        <NavigationBreadcrumbs
-                            title="Bills"
-                            breadcrumbs={[{title: 'Account', url: '/consumptions'}]}
-                        />
-                    }
-                    extra={<Placeholder />}
-                />
-            </div>
-            <ResponsiveLayout>
-                <Stack space={16}>
-                    <Text2 medium>Documentation:</Text2>
-                    <Text1 regular>
-                        The header is optional in order to allow webviews to delegate the header visualization
-                        to the surrounding native app. For example, in Novum App, the Start tab's greeting is
-                        rendered nativelly in the apps and via web in desktop.
-                    </Text1>
-                    {extraSideBySideCheckbox}
-                </Stack>
-            </ResponsiveLayout>
-        </Stack>
-    );
+Default.args = {
+    withHeader: true,
+    pretitle: 'Your last bill',
+    title: 'December bill is now available',
+    description: 'This is a description',
+    small: false,
+    isInverse: true,
+    withBreadcrumbs: true,
+    noPaddingY: false,
+    withExtraContent: true,
+    sideBySideExtraOnDesktop: true,
+    bleed: false,
 };
-
-NoHeader.storyName = 'Header layout with no header';
+Default.argTypes = {
+    pretitle: {if: {arg: 'withHeader'}},
+    title: {if: {arg: 'withHeader'}},
+    description: {if: {arg: 'withHeader'}},
+    small: {if: {arg: 'withHeader'}},
+    sideBySideExtraOnDesktop: {if: {arg: 'withExtraContent'}},
+    bleed: {if: {arg: 'withExtraContent'}},
+};
 
 export const RichTexts: StoryComponent = () => {
     const filler = ' - more text'.repeat(20);
     return (
-        <div data-testid="header-layout">
-            <HeaderLayout
-                header={
-                    <Header
-                        pretitle={{
-                            text: `Pretitle (truncated to one line) ${filler}}`,
-                            truncate: true,
-                        }}
-                        title="Title is always a plain string"
-                    />
-                }
-            />
-        </div>
+        <HeaderLayout
+            dataAttributes={{testid: 'header-layout'}}
+            header={
+                <Header
+                    pretitle={{
+                        text: `Pretitle (truncated to one line) ${filler}}`,
+                        truncate: true,
+                    }}
+                    title="Title is always a plain string"
+                />
+            }
+        />
     );
 };