+++ src/index.tsx
@@ -61,7 +61,16 @@ export {default as HorizontalScroll} from './horizontal-scroll';
 export {default as HighlightedCard} from './highlighted-card';
 export {default as Stepper} from './stepper';
 export {default as ProgressBar} from './progress-bar';
-export {MediaCard, DataCard, SnapCard, DisplayDataCard, DisplayMediaCard, PosterCard} from './card';
+export {
+    MediaCard,
+    DataCard,
+    SnapCard,
+    DisplayDataCard,
+    DisplayMediaCard,
+    PosterCard,
+    NakedCard,
+    SmallNakedCard,
+} from './card';
 export {default as Hero} from './hero';
 export {default as Divider} from './divider';
 export {default as Menu} from './menu';